chore: lsb separation
